Refac: move prediction utils to lightning.prediction.lvae
