+++ src/careamics/lightning/prediction/lvae/lvae_convert_outputs.py
