+++ src/careamics/lightning/prediction/lvae/lvae_prediction.py
