+++ src/careamics/lightning/prediction/lvae/lvae_stitch_prediction.py
