+++ tests/prediction_utils/test_lvae_prediction.py
@@ -8,12 +8,12 @@ from careamics.compat.config.data.tile_information import TileInformation
 from careamics.compat.dataset import InMemoryTiledPredDataset
 from careamics.compat.dataset.tiling.collate_tiles import collate_tiles
 from careamics.compat.prediction_utils import convert_outputs
-from careamics.models.lvae.likelihoods import GaussianLikelihood
-from careamics.models.lvae.lvae import LadderVAE
-from careamics.prediction_utils.lvae_prediction import (
+from careamics.lightning.prediction.lvae.lvae_prediction import (
     lvae_predict_mmse_tiled_batch,
     lvae_predict_tiled_batch,
 )
+from careamics.models.lvae.likelihoods import GaussianLikelihood
+from careamics.models.lvae.lvae import LadderVAE
 
 # TODO: Test with mock LCMultiChDloader
